Fix bswap_64 for x86-64
Use the correct type name for the 64-bit integer type.

BUG= http://code.google.com/p/chromium/issues/detail?id=133889
TEST= locally tested it fixes run_byteswap64_test

R=bradchen@google.com

Review URL: https://chromiumcodereview.appspot.com/10702071

+++ sysdeps/x86_64/bits/byteswap.h
@@ -110,7 +110,7 @@
 # if __WORDSIZE == 64
 #  define __bswap_64(x) \
      (__extension__							      \
-      ({ register unsigned long __v, __x = (x);				      \
+      ({ register unsigned long long __v, __x = (x);			      \
 	 if (__builtin_constant_p (__x))				      \
 	   __v = __bswap_constant_64 (__x);				      \
 	 else								      \
